+++ frontend/src/papaparse.d.ts
@@ -1,0 +1,1 @@
+declare module 'papaparse';
